main/requirements.txt
Flask
Flask-SQLAlchemy
SQLAlchemy
Flask-Migrate
Flask-Cors
requests
mysqlclient
kafka-python